fix, Jumping from *SymflowerTest.java file to the implementation
The switch statement in symflower-alternative-file has two cases
*.java and *SymflowerTest.java. Since the first check subsumes
the second, that one is dead code, so we fail to jump back to the
implementation. Let's switch the order, which also matches the semantic
order of the Go cases. Also add a comment to explain why we don't
use lexical order.

This reverts commit c495c23 (refactor, Sort, 2022-01-08).