Embelish the documentation with Asciinema screencasts
Add three screencasts that showcase three different features:

1. unit test generation
2. user mode (shortcuts)
3. unit test skeleton generation

add them next to their respective documentation entries.  Except for
the first one, because we use that as opener.  It seems stronger than
the static image of Copy.java, let's just move that to "Quick Start"
(which fits because it uses the same Copy.java).